+++ frontend/src/components/map-interface/MapFilters.tsx
@@ -11,7 +11,7 @@ import {
   X,
 } from "lucide-react";
 import { useEffect, useState } from "react";
-import { Button } from "./ui/button";
+import { Button } from "../ui/button";
 import { cn } from "@/lib/utils";
 
 const MapFilters = () => {
